requirements.txt
awscli-cwlogs==1.4.0
boto3>=1.4.4
six>=1.10.0
